CI fix

+++ Sources/NinetyNineSwiftProblems/Random.swift
@@ -3,7 +3,7 @@ import Foundation
 public func randomInt(under limit: Int) -> Int {
     #if os(Linux)
         srandom(UInt32(time(nil)))
-        return UInt32(random() % limit)
+        return random() % limit
     #else
         return Int(arc4random_uniform(UInt32(limit)))
     #endif
